closes #5
Habemus time! It's not as beauty as I'd like it to be, but well, it's
the first attempt. Now we can use times as this:

1.hour and 5.minutes
1.h & 5.m & 3.s

Besides, we can add two times:

(1.h & 1.m & 1.s) and (2.h & 2.m & 2.s)

Finally, time can be normalized. What's that? Well, it means that
90.seconds should be 1.minute & 30.seconds, but some trainers do like
to work with "extra" minutes, don't they?